Mount the whole rootfs of the Node instead of a single folder
When doing a hardlink for the containerd.sock it won't work the
two folders are on different filesystems. Symlinks will work but only
if we have a full view of the Node's root filesystem. The security
of this is mostly the same since the acccess to containerd.sock is
already giving us full access to the Node.